allow file paths in aggregate_file script and rename plot and widget parent class
