+++ src/autogluon_dashboard/plotting/errored_datasets.py
@@ -6,7 +6,7 @@ import pandas
 from autogluon_dashboard.constants.df_constants import DATASET
 
 from ..utils.dataset_utils import get_df_filter_by_framework, get_sorted_names_from_col
-from .all_plots import Plot
+from .plot import Plot
 
 ERRORED_DATASETS = "Errored Datasets"
