+++ src/autogluon_dashboard/plotting/metrics_per_datasets.py
@@ -4,7 +4,7 @@ import hvplot
 import pandas
 
 from ..utils.dataset_utils import get_df_filter_by_dataset
-from .all_plots import Plot
+from .plot import Plot
 
 
 class MetricsPlotPerDataset(Plot):
