+++ src/autogluon_dashboard/plotting/top5_all_datasets.py
@@ -4,7 +4,7 @@ import hvplot
 import pandas
 
 from ..utils.dataset_utils import get_top5_performers
-from .all_plots import Plot
+from .plot import Plot
 
 
 class Top5AllDatasets(Plot):
